feat(full-update-status): Create endpoint to fully update status.
Add 'fullUpdate' method to StatusService. Implement the method in StatusServiceImpl checking that status with the id passed in exists, normalising & finally saving it. Add @Transactional annotation to avoid unexpected behaviours as this method uses two db operations. Add PUT fullUpdateStatus method in StatusController with an 'id' as path variable, and the body of the statusDto with the updated values.