feat(partial-update-status): Create endpoint to partially update status.
Add 'partialUpdate' method to StatusService. Implement the method in StatusServiceImpl checking that the user retrieved by the id is not null, setting the new values & finally saving it. Add @Transactional annotation to avoid unexpected behaviours as this method uses two db operations. Add PATCH partialUpdateStatus method in StatusController with an 'id' as path variable, and the body of the statusUpdateRequestDto with the updated values. Create separate entities & dto for partial update as some values can be null. Create extension functions for such data classes.